refactor(gateway): extract auto-continue helpers from run.py (slice 8 of #54962)
